+++ database/models/__init__.py
@@ -1,1 +1,2 @@
 from .tyre_model import TyreModel
+from .tyre_impression import TyreImpression