+++ database/models/data_types/__init__.py
@@ -1,0 +1,1 @@
+from .tyre_impression_status import TyreImpressionStatus